Add a couple of requirements
My environment turned out to be kind of antique, because Querki is *deeply* antique, so I discovered the hard way a couple of upgrades I had to do. Trivial stuff, but let's say it out loud for anyone else who might be in a similar boat.

After dealing with those, all runs smooth as silk!